Implemented drag and drop for PDF columns fragment, refactoring
