+++ app/src/main/java/co/smartreceipts/android/settings/widget/editors/columns/PDFColumnsListFragment.java
@@ -6,33 +6,27 @@ import android.support.v7.app.ActionBar;
 import javax.inject.Inject;
 
 import co.smartreceipts.android.R;
+import co.smartreceipts.android.model.Column;
+import co.smartreceipts.android.model.Receipt;
 import co.smartreceipts.android.model.impl.columns.receipts.ReceiptColumnDefinitions;
-import co.smartreceipts.android.persistence.database.controllers.impl.ColumnTableController;
+import co.smartreceipts.android.persistence.database.controllers.TableController;
 import co.smartreceipts.android.persistence.database.controllers.impl.PDFTableController;
+import co.smartreceipts.android.persistence.database.tables.ordering.OrderingPreferencesManager;
 import dagger.android.support.AndroidSupportInjection;
 
-public class PDFColumnsListFragment extends ColumnsListFragment1 {
+public class PDFColumnsListFragment extends ColumnsListFragment {
 
     public static String TAG = "PDFColumnsListFragment";
-
 
     @Inject
     ReceiptColumnDefinitions receiptColumnDefinitions;
     @Inject
     PDFTableController pdfTableController;
+    @Inject
+    OrderingPreferencesManager orderingPreferencesManager;
 
-    public static ColumnsListFragment1 newInstance() {
+    public static PDFColumnsListFragment newInstance() {
         return new PDFColumnsListFragment();
-    }
-
-    @Override
-    protected ColumnTableController getColumnTableController() {
-        return pdfTableController;
-    }
-
-    @Override
-    protected ReceiptColumnDefinitions getReceiptColumnDefinitions() {
-        return receiptColumnDefinitions;
     }
 
     @Override
@@ -48,6 +42,22 @@ public class PDFColumnsListFragment extends ColumnsListFragment1 {
         if (actionBar != null) {
             getSupportActionBar().setTitle(R.string.menu_main_pdf);
         }
+    }
+
+    @Override
+    protected TableController<Column<Receipt>> getTableController() {
+        return pdfTableController;
+    }
+
+    @Override
+    protected ReceiptColumnDefinitions getReceiptColumnDefinitions() {
+        return receiptColumnDefinitions;
+    }
+
+    @Override
+    protected void saveTableOrdering() {
+        super.saveTableOrdering();
+        orderingPreferencesManager.savePdfColumnsTableOrdering();
     }
 
 }
