--- app/src/main/res/layout/settings_column_card_item.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-	style="@style/Widget.SmartReceipts.LinearLayout.Card" >
-
-	<TextView
-	    android:id="@android:id/title"
-	    android:layout_width="0dp"
-	    android:layout_height="wrap_content"
-	    android:layout_gravity="center_vertical"
-	    android:textColor="@color/card_name"
-	    android:textSize="@dimen/card_name_size"
-	    android:textStyle="italic"
-	    android:layout_marginLeft="5dp"
-		android:layout_marginRight="5dp"
-		android:layout_weight="1" />
-		
-	<Spinner
-		android:id="@+id/column_spinner"
-		android:layout_width="0dp"
-		android:layout_height="wrap_content"
-		android:layout_gravity="center_vertical"
-		android:prompt="@string/dialog_custom_csv_spinner"
-		android:layout_marginLeft="5dp"
-		android:layout_marginRight="5dp"
-		android:layout_weight="2" />
-	   
-	<!-- 
-	<View style="@style/Widget.View.Divider.Vertical" />
-	       
-	<ImageView
-		android:id="@+id/delete"
-		android:layout_width="wrap_content"
-		android:layout_height="match_parent"
-		android:src="@drawable/ic_action_discard"
-		android:padding="4dp"
-		android:contentDescription="@string/delete"
-		android:background="@drawable/default_selector" />
-	 -->
-           
-</LinearLayout>
